fastapi-web-macos122/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
Pillow